Expanded the test to cover verifying more data is deleted, added option to delete your account just after initial creation, added flash message to alert user their account has been deleted

+++ src/Http/Action/Profile/ProcessDeleteAction.php
@@ -41,9 +41,23 @@ final class ProcessDeleteAction
     public function __invoke(HttpFoundation\Request $request): HttpFoundation\Response
     {
         $user = User::find($this->authentication->user()->getId());
+
+        if (!$user) {
+            $url = $this->urlGenerator->generate('login');
+
+            return new HttpFoundation\RedirectResponse($url);
+        }
+
         $user->delete();
         $this->authentication->logout();
         $url = $this->urlGenerator->generate('homepage');
+
+        // Set flash message acknowledging that your account has been deleted
+        $request->getSession()->set('flash', [
+            'type'  => 'success',
+            'short' => 'Account Deleted',
+            'ext'   => 'Your OpenCFP account here has been deleted',
+        ]);
 
         return new HttpFoundation\RedirectResponse($url);
     }
